tweaked I2CDev to get rid of some compilation warnings and updated todo

+++ Tilty Software/Arduino/Libraries/I2Cdev/I2Cdev.cpp
@@ -236,7 +236,7 @@ int8_t I2Cdev::readBytes(uint8_t devAddr, uint8_t regAddr, uint8_t length, uint8
                 Wire.send(regAddr);
                 Wire.endTransmission();
                 Wire.beginTransmission(devAddr);
-                Wire.requestFrom((int)devAddr, (int)min(length - k, BUFFER_LENGTH));
+                Wire.requestFrom(devAddr, min(length - k, BUFFER_LENGTH));
 
                 for (; Wire.available() && (timeout == 0 || millis() - t1 < timeout); count++) {
                     data[count] = Wire.receive();
